+++ crates/tswn_py/src/wrapper/rc4.rs
@@ -1,4 +1,4 @@
-use pyo3::{PyResult, exceptions::PyValueError, pyclass, pymethods};
+use pyo3::{exceptions::PyValueError, pyclass, pymethods, PyResult};
 
 use tswn_core::rc4::{RC4, VAL_LEN};